topcat: use generics throughout
Eliminate or suppress all compile-time warnings at java 8.
There are a lot of lines of source code changed, but most of them
do not result in any changes that don't get erased at compile time.
There should be no behaviour changes.

SuppressWarnings is used in quite a few places, notably when dealing
with classes like JComboBox that have acquired generics since java 6.

+++ topcat/src/main/uk/ac/starlink/topcat/TableViewerWindow.java
@@ -656,7 +656,7 @@ public class TableViewerWindow extends AuxWindow {
     public void setSelection( RowSubset rset ) {
         rowSelectionModel_.setValueIsAdjusting( true );
         rowSelectionModel_.clearSelection();
-        int nrow = (int) viewModel_.getRowCount();
+        int nrow = viewModel_.getRowCount();
         int[] rowMap = viewModel_.getRowMap();
         for ( int irow = 0; irow < nrow; irow++ ) {
             long jrow = rowMap == null ? (long) irow
